// Web App/src/front end/cars-web-app/src/app/features/mechanics/models/mechanics.models.ts
export interface MechanicDTO{
    name: string;
    cnp: string;
    address: string;
    experience: number;
    favouriteCarBrand: string;
}

export interface MechanicIDDTO{
    id: number;
    name: string;
    cnp: string;
    address: string;
    experience: number;
    favouriteCarBrand: string;
}